LayoutTests/media/W3C/audio/canPlayType/canPlayType_application_octet_stream-expected.txt

audio.canPlayType() - application/octet-stream must always return the empty string

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS true is true

TEST COMPLETE
spec reference
